refactor(ui): split App.tsx into 10 per-tab components
App.tsx drops from 1,156 lines to a ~115-line shell (routing, sidebar,
header, tab switch). Each tab is its own component under src/components/tabs/,
reading shared data from context and owning its single-tab polls:

- LiveTab, AgentsTab, WorkflowsTab, TrendsTab, ModelsTab, InsightsTab,
  WorkspaceTab, AiTab, SessionsTab, SettingsTab

TrendsTab composes the three new organisms; insight-enabled Sections now use
the aiProps() helper instead of the repeated five-prop bundle. Global toasts/
analytics moved into useDashboardNotifications. Behavior unchanged.
