+++ src/components/tabs/LiveTab/types.ts
@@ -1,0 +1,5 @@
+import type { Limits } from '@/hooks/useLimits';
+
+export interface LiveTabProps {
+  limits: Limits;
+}
