+++ src/components/tabs/SettingsTab/types.ts
@@ -1,0 +1,6 @@
+import type { Limits } from '@/hooks/useLimits';
+
+export interface SettingsTabProps {
+  limits: Limits;
+  onChangeLimits: (l: Limits) => void;
+}
